+++ PyStageLinQ/PyStageLinQ.py
@@ -8,6 +8,7 @@ import socket
 import time
 import asyncio
 import logging
+import psutil
 from typing import Callable
 
 from . import Device
@@ -63,11 +64,13 @@ class PyStageLinQ:
 
         self.device_list = Device.DeviceList()
 
+        self.ip = []
         if ip is None:
-            interfaces = socket.getaddrinfo(
-                host=socket.gethostname(), port=None, family=socket.AF_INET
-            )
-            self.ip = [ip[-1][0] for ip in interfaces]
+            interfaces = psutil.net_if_addrs()
+            for interface in interfaces.items():
+                for interface_address in interface[1]:
+                    if socket.AF_INET == interface_address.family:
+                        self.ip.append(interface_address.address)
 
         else:
             self.ip = [ip]
@@ -84,12 +87,15 @@ class PyStageLinQ:
 
         self.new_device_found_callback = new_device_found_callback
 
+        logger.debug(f"Initialized!")
+
     def start_standalone(self):
         """
         Function for starting PyStageLinq in standalone mode. In this mode this function will not return, but
         instead start asyncio and que up tasks. This function shall be called once the PyStageLinQ object has been
         initialized.
         """
+        logger.debug(f"Starting in standalone mode.")
         asyncio.run(self._start_stagelinq(standalone=True))
 
     def start(self):
@@ -98,9 +104,11 @@ class PyStageLinQ:
         tasks and then return to its caller. It is then up to the called to make sure that the asyncio tasks gets
         time to execute. This function shall be called once the PyStageLinQ object has been initialized.
         """
+        logger.debug(f"Starting in asyncio mode.")
         self._start_stagelinq()
 
     def _stop(self):
+        logger.info(f"Stop requested, trying graceful shutdown")
         try:
             discovery = StageLinQDiscovery()
             discovery_info = self.discovery_info
@@ -108,6 +116,7 @@ class PyStageLinQ:
             discovery_frame = discovery.encode_frame(discovery_info)
 
             self._send_discovery_frame(discovery_frame)
+            logger.info(f"Gracefully shutdown complete")
         except:
             logger.debug('Could not send "EXIT" discovery frame during shutdown')
 
@@ -127,10 +136,11 @@ class PyStageLinQ:
                         ("255.255.255.255", self.StageLinQ_discovery_port),
                     )
                 except PermissionError:
-                    raise PermissionError(
+                    logger.warning(
                         f"Cannot send message on interface {ip}, "
                         f"this error could be due to that there is no network interface set up with this IP range"
                     )
+                    raise PermissionError
 
     def get_loop_condition(self) -> bool:
         return self._loopcondition
@@ -139,6 +149,8 @@ class PyStageLinQ:
         """
         This function is used to find StageLinQ device announcements.
         """
+        logger.info(f"Trying to discover StageLinQ devices.")
+
         # Local Constants
         discover_buffer_size = 8192
 
@@ -146,14 +158,21 @@ class PyStageLinQ:
         discover_socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
         try:
             discover_socket.bind(
-                (host_ip, self.StageLinQ_discovery_port)
+                ("255.255.255.255", self.StageLinQ_discovery_port)
             )  # bind socket StageLinQ interface
         except:
             # Cannot bind to socket, check if IP is correct and link is up
+            logger.warning(
+                f"Cannot bind to IP socket: {host_ip} on port {self.StageLinQ_discovery_port}"
+            )
             return PyStageLinQError.CANNOTBINDSOCKET
         discover_socket.setblocking(False)
 
         loop_timeout = time.time() + timeout
+
+        logger.debug(
+            f"Socket bound to IP {host_ip} and Port {self.StageLinQ_discovery_port} successfully. Starting to look for discovery frames"
+        )
 
         while self.get_loop_condition():
             await asyncio.sleep(0.1)
@@ -183,13 +202,16 @@ class PyStageLinQ:
                 )
                 if device_registered is True:
                     continue
+
+                logger.info(f"Found new StageLinq device at IP {device_ip}")
                 await self._register_new_device(discovery_frame, device_ip)
 
             if time.time() > loop_timeout:
                 # No devices found within timeout
+                logger.info(
+                    "No discovery frames found on {host_ip} last {timeout} seconds."
+                )
                 return PyStageLinQError.DISCOVERYTIMEOUT
-
-        logger.info(f"No discovery frames found on {host_ip} last {timeout} seconds.")
 
     async def _register_new_device(self, discovery_frame, ip):
         stagelinq_device = StageLinQService(ip, discovery_frame, self.OwnToken, None)
@@ -197,6 +219,7 @@ class PyStageLinQ:
         for task in service_tasks:
             self.tasks.add(task)
         self.device_list.register_device(stagelinq_device)
+        logging.debug(f"Device found! Name: {stagelinq_device.device_name}")
         await stagelinq_device.wait_for_services(timeout=1)
         if self.new_device_found_callback is not None:
             self.new_device_found_callback(
@@ -216,6 +239,7 @@ class PyStageLinQ:
                 :param data_available_callback: Callback for when data is available from StageLinQ device
         """
         if state_map_service.service != "StateMap":
+            logger.warning(f'{state_map_service.service} is not of type "Statemap"')
             return PyStageLinQError.SERVICENOTRECOGNIZED
 
         # Defer task creation to avoid blocking the calling function
@@ -235,6 +259,7 @@ class PyStageLinQ:
         await state_map.subscribe(self.OwnToken)
 
         self.tasks.add(state_map.get_task())
+        logger.debug(f"Subscription to StateMap successful.")
 
     async def _start_stagelinq(self, standalone=False):
         # Start the initial tasks of the library
@@ -250,12 +275,17 @@ class PyStageLinQ:
             for task in self.tasks.copy():
                 if task.done():
                     if task.exception() is not None:
+                        logger.warning(
+                            f"Exception: {task.exception()} occured in task: {task.get_coro()}, stopping "
+                            f"PyStageLinQ."
+                        )
                         self.stop()
                         raise task.exception()
                     return
             await asyncio.sleep(1)
 
     def _stop_all_tasks(self):
+        logger.info(f"Stop all PyStageLinQ AsyncIO tasks requested")
         for task in self.tasks.copy():
             task.cancel()